pyproject.toml
[tool.poetry]
name = "human-text-input-clustering"
version = "0.1.0"
description = ""
authors = ["mihas <mihasK@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "human_text_input_clustering"}]

[tool.poetry.dependencies]
python = "^3.11"
sklearn-pandas = "^2.2.0"
jupyter = "^1.0.0"
dedupe = "^2.0.23"
rapidfuzz = "^3.1.1"
icecream = "^2.1.3"
unidecode = "^1.3.6"
dash = "^2.11.1"
loguru = "^0.7.0"
dash-bootstrap-components = "^1.4.1"
snoop = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
